+++ scripts/ci/boundary_maps/receipts_verifier.json
@@ -8,6 +8,7 @@
     "aragora/gauntlet/receipt_models.py",
     "aragora/gauntlet/receipt.py",
     "aragora/receipts",
+    "aragora/gauntlet/odr_jcs.py",
     "aragora/gauntlet/odr_export.py",
     "aragora/gauntlet/odr_signing.py",
     "aragora/gauntlet/odr_verify.py",
@@ -30,6 +31,7 @@
     "aragora.gauntlet.receipt_models",
     "aragora.gauntlet.receipt",
     "aragora.receipts",
+    "aragora.gauntlet.odr_jcs",
     "aragora.gauntlet.odr_export",
     "aragora.gauntlet.odr_signing",
     "aragora.gauntlet.odr_verify"
